Add proxy support
Allow ES-Head to be used when the ES servers are in a restricted area
for security reasons or network configuration. Edit index.html and
change `proxy: false` to your own proxy. The proxy will be called with
the 'uri' parameter, see proxy.php as an example. Your proxy can also for
example handle authorization. Of course your proxy must have access to
the ES servers.